+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Fix up the `doc_width` calculation for `CommandCall` nodes.
 - Ensure parameters inside a lambda literal when there are no parentheses are grouped.
 - Ensure when converting an `ArrayLiteral` to a `QWords` that the strings do not contain `[`.
+- Stop looking for parent `Command` or `CommandCall` nodes in blocks once you hit `Statements`.
 
 ## [0.1.0] - 2021-11-16